+++ tests/correctness/targets/CppCompilationWithTargetDeps/Input/test.cpp
@@ -3,10 +3,11 @@
 #include <somepath/test1.h>
 #include <generatedpath/test3.h>
 #include <somepath/test2.h>
+#include <somepath/test4.h>
 
 int main(int argc, char **argv)
 {
 	printf("Test program\n");
-	printf("Got string: <" STRING1 STRING2 ">\n");
+	printf("Got string: <" STRING1 STRING2 STRING3 STRING4 ">\n");
 	return 0;
 }